pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openai-cost-tracker"
version = "0.0.1"
description = "A Python package for tracking and estimating costs when using OpenAI API"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Timur Zhilyaev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "httpx",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "flake8",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/madeinmo/openai-cost-tracker"
Repository = "https://github.com/madeinmo/openai-cost-tracker"
Documentation = "https://github.com/madeinmo/openai-cost-tracker"

[tool.setuptools]
packages = ["openai_cost_tracker"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
